ml/requirements.txt
rapidfuzz>=3.9.0
pydantic>=2.9.0
numpy>=1.26.0
pillow>=10.4.0
pyyaml>=6.0.2
